[63강] : Feature 및 State로 컴포넌트 분리
컴포넌트를 더 작은 단위로 분리하는 방법을 공부했습니다.
