+++ practice/01-starting-project/src/App.jsx
@@ -1,56 +1,23 @@
 import componentsImg from './assets/Components.png';
-import { CORE_CONCEPTS } from './data.js';
 
 // 컴포넌트 분리
 import Header from'./components/Header/Header.jsx';
-import CoreConcept from'./components/CoreConcept.jsx';
-import TabButton from'./components/TabButton.jsx';
+import CoreConcepts from'./components/CoreConcepts.jsx';
 
-import { useState } from 'react';
-import { EXAMPLES } from './data.js';
+import { useState, Fragment } from 'react';
+import Examples from './components/Examples.jsx';
 
 function App() {
 
-  let [selectedTopic, setSelectedTopic] = useState();
-
-  function handleSelect(selectedButton) {
-    setSelectedTopic(selectedButton);
-    console.log(selectedButton);
-  } 
-
   return (
-    <div>
+    <Fragment>
       <Header />
       <main>
-        <section id="core-concepts">
-          <h2>Core Concepts</h2>
-          <ul>
-            {CORE_CONCEPTS.map((conceptItem) => <CoreConcept key={conceptItem.title} {...conceptItem}/> )}
-          </ul>
-        </section>
-        <section id="examples">
-          <h2>Examples</h2>
-          <menu>
-            <TabButton isSelected = {selectedTopic === 'components'} onSelect={() => handleSelect('components')}>Components</TabButton>
-            <TabButton isSelected = {selectedTopic === 'jsx'} onSelect={() => handleSelect('jsx')}>JSX</TabButton>
-            <TabButton isSelected = {selectedTopic === 'props'} onSelect={() => handleSelect('props')}>Props</TabButton>
-            <TabButton isSelected = {selectedTopic === 'state'} onSelect={() => handleSelect('state')}>State</TabButton>
-          </menu>
-          {!selectedTopic ? (<p>Please select a topic</p>) :
-          ( <div id="tab-content">
-            <h3>{EXAMPLES[selectedTopic].title}</h3>
-            <p>{EXAMPLES[selectedTopic].description}</p>
-            <pre>
-              <code>
-                {EXAMPLES[selectedTopic].code}
-              </code>
-            </pre>
-          </div>
-          )}
-        </section>
+        <CoreConcepts/>    
+        <Examples/>    
         <h2>Time to get started!</h2>
       </main>
-    </div>
+    </Fragment>
   );
 }
