+++ practice/01-starting-project/src/components/CoreConcepts.jsx
@@ -1,0 +1,13 @@
+import CoreConcept from "./CoreConcept"
+import { CORE_CONCEPTS } from "../data"
+
+export default function CoreConcepts({}) {
+    return(
+        <section id="core-concepts">
+          <h2>Core Concepts</h2>
+          <ul>
+            {CORE_CONCEPTS.map((conceptItem) => <CoreConcept key={conceptItem.title} {...conceptItem}/> )}
+          </ul>
+        </section>
+    )
+}
